ai-service/requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
langchain
langchain-community
langchain-cerebras
langchain-huggingface
huggingface-hub
cerebras_cloud_sdk
faiss-cpu
pypdf
pandas
python-dotenv